MAINT: Move uarray into scipy._lib
