+++ scipy/_lib/setup.py
@@ -43,6 +43,8 @@ def configuration(parent_package='',top_path=None):
                          depends=depends,
                          include_dirs=[include_dir])
 
+    config.add_subpackage('uarray')
+
     return config
