+++ scipy/_lib/uarray/setup.py
@@ -1,7 +1,7 @@
 def configuration(parent_package='', top_path=None):
     from numpy.distutils.misc_util import Configuration
 
-    config = Configuration('_uarray', parent_package, top_path)
+    config = Configuration('uarray', parent_package, top_path)
     config.add_data_files('LICENSE')
     return config
