+++ scipy/fft/_basic.py
@@ -1,4 +1,4 @@
-from scipy._uarray import generate_multimethod, Dispatchable
+from scipy._lib.uarray import generate_multimethod, Dispatchable
 import numpy as np
 
 def _x_replacer(args, kwargs, dispatchables):
